founding-corpus/avalon_scraper/requirements.txt
httpx[http2]==0.27.0
beautifulsoup4==4.12.3
html5lib==1.1
pdfminer.six==20231228
pandas==2.2.2
pyarrow==16.1.0
python-dateutil==2.9.0.post0
xxhash==3.4.1
orjson==3.10.7
robotexclusionrulesparser==1.7.1
tenacity==8.2.3
tqdm==4.66.4